requirements.txt
requests==2.32.5
python-dotenv==1.2.2
PyYAML==6.0.3
pandas==3.0.1
pyarrow==23.0.1

#Big Data / Spark (For M3)
#pyspark==3.5.3